move message logic out of extension/backtrack functions and put in reducer instead. also pass in puzzle.
